Add type cast

+++ src/motion_decision.cpp
@@ -300,8 +300,8 @@ std::pair<float, float> MotionDecision::sim_by_uniform_circluar_motion(const flo
 
 void MotionDecision::publish_cmd_vel(geometry_msgs::Twist cmd_vel)
 {
-  cmd_vel.linear.x = 0.0 < cmd_vel.linear.x ? std::min(cmd_vel.linear.x, params_.max_speed) : std::max(cmd_vel.linear.x, -params_.max_speed);
-  cmd_vel.angular.z = 0.0 < cmd_vel.angular.z ? std::min(cmd_vel.angular.z, params_.max_yawrate) : std::max(cmd_vel.angular.z, -params_.max_yawrate);
+  cmd_vel.linear.x = 0.0 < cmd_vel.linear.x ? std::min(cmd_vel.linear.x, static_cast<double>(params_.max_speed)) : std::max(cmd_vel.linear.x, static_cast<double>(-params_.max_speed));
+  cmd_vel.angular.z = 0.0 < cmd_vel.angular.z ? std::min(cmd_vel.angular.z, static_cast<double>(params_.max_yawrate)) : std::max(cmd_vel.angular.z, -static_cast<double>(params_.max_yawrate));
 
   if (flags_.emergency_stop || mode_.first == "stop")
     cmd_vel = geometry_msgs::Twist();
